+++ pyproject.toml
@@ -33,11 +33,24 @@ numpy = ">=1.26.2"
 matplotlib = ">=3.8.1"
 setuptools = ">=69.0.0"
 rtree = "^1.4.0"
+isort = "^6.0.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.3"
 pytest-cov = "^4.1.0"
 pylint = "^3.0.2"
+pre-commit = "^3.0.0"
+
+[tool.isort]
+profile = "black"
+line_length = 88
+multi_line_output = 3
+include_trailing_comma = true
+force_grid_wrap = 0
+use_parentheses = true
+ensure_newline_before_comments = true
+known_first_party = ["geo_adjacency"]
+known_third_party = ["shapely", "scipy", "numpy", "matplotlib", "rtree", "pytest"]
 
 [build-system]
 requires = ["poetry-core"]
